feat(jr): implement home component with topic cards and routing
