+++ angularApp/src/app/app.routes.ts
@@ -1,3 +1,8 @@
 import { Routes } from '@angular/router';
+import { HomeComponent } from './home/home.component';
+import { ReactiveComponent } from './reactive/reactive.component';
 
-export const routes: Routes = [];
+export const routes: Routes = [
+  { path: '', component: HomeComponent },
+  { path: 'reactive', component: ReactiveComponent },
+];
