+++ angularApp/src/app/home/home.component.scss
@@ -1,0 +1,14 @@
+/* home component — layout uses shared design system from styles.scss */
+
+.topic-card {
+  display: block;
+  text-decoration: none;
+  color: inherit;
+  cursor: pointer;
+  transition: transform 0.15s ease, box-shadow 0.15s ease;
+
+  &:hover {
+    transform: translateY(-2px);
+    box-shadow: 0 6px 24px rgba(0, 0, 0, 0.12);
+  }
+}
